Move rotate button up one level and make preview field scrollable for mobile
