+++ src/components/attack/Preview.css
@@ -1,9 +1,6 @@
 .preview-wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  margin-top: 15px;
+  max-width: 100%;
+  overflow-x: auto;
 }
 
 .preview-fields-wrapper {
